# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: B2B AI Marketing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/campaigns:
    get:
      operationId: getCampaigns
      tags: [analytics]
      summary: List campaigns with metrics
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dateRange
          in: query
          schema:
            type: ["string", "null"]
            enum: [today, week, month, quarter, null]
      responses:
        "200":
          description: Campaign list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
  /analytics/breakdown:
    get:
      operationId: getAnalyticsBreakdown
      tags: [analytics]
      summary: Get analytics broken down by client, department, or role
      parameters:
        - name: groupBy
          in: query
          required: true
          schema:
            type: string
            enum: [client, department, role]
      responses:
        "200":
          description: Analytics breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BreakdownEntry"
  /analytics/trends:
    get:
      operationId: getAnalyticsTrends
      tags: [analytics]
      summary: Get trend data for charts (time series)
      parameters:
        - name: metric
          in: query
          required: true
          schema:
            type: string
            enum: [impressions, ctr, leads, seats, utilization]
        - name: days
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Trend data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"

  # ── Clients ─────────────────────────────────────────────────────────────────
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: List all client organizations
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  # ── Departments ──────────────────────────────────────────────────────────────
  /departments:
    get:
      operationId: getDepartments
      tags: [departments]
      summary: List departments
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Department list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"

  # ── Users ────────────────────────────────────────────────────────────────────
...
  /audit-logs:
    get:
      operationId: getAuditLogs
      tags: [audit]
      summary: List audit log entries
      parameters:
        - name: entityType
          in: query
          schema:
            type: ["string", "null"]
            enum: [license, seat, user, role, client, null]
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Audit log list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"

  # ── Activity Feed ─────────────────────────────────────────────────────────────
  /activity-feed:
    get:
      operationId: getActivityFeed
      tags: [activity]
      summary: Get live activity feed events
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Activity events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
components:
  schemas:
...
    Campaign:
      type: object
      required: [id, name, clientId, clientName, impressions, ctr, leads, status, startDate]
      properties:
        id:
          type: integer
        name:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        impressions:
          type: integer
        ctr:
          type: number
        leads:
          type: integer
        leadQuality:
          type: ["string", "null"]
          enum: [high, medium, low, null]
        status:
          type: string
          enum: [active, paused, completed, draft]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
    BreakdownEntry:
      type: object
      required: [label, impressions, ctr, leads, seats, utilization]
      properties:
        label:
          type: string
        impressions:
          type: integer
        ctr:
          type: number
        leads:
          type: integer
        seats:
          type: integer
        utilization:
          type: number
    TrendPoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
        value:
          type: number

    # ── Client Schemas ─────────────────────────────────────────────────────────
    Client:
      type: object
      required: [id, name, status, totalLicenses, activeSeats, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, suspended]
        totalLicenses:
          type: integer
        activeSeats:
          type: integer
        inactiveSeats:
          type: integer
        utilizationRate:
          type: ["number", "null"]
        createdAt:
          type: string
        contactEmail:
          type: ["string", "null"]
    ClientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        industry:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, suspended]
    ClientUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: [active, inactive, suspended, null]

    # ── Department Schemas ──────────────────────────────────────────────────────
    Department:
      type: object
      required: [id, name, clientId, memberCount, activeSeatCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        memberCount:
          type: integer
        activeSeatCount:
          type: integer
        createdAt:
          type: string
    DepartmentInput:
      type: object
      required: [name, clientId]
      properties:
        name:
          type: string
        clientId:
          type: integer

    # ── User Schemas ───────────────────────────────────────────────────────────
...
    AuditLog:
      type: object
      required: [id, action, entityType, entityId, performedBy, timestamp]
      properties:
        id:
          type: integer
        action:
          type: string
        entityType:
          type: string
          enum: [license, seat, user, role, client]
        entityId:
          type: integer
        entityName:
          type: ["string", "null"]
        performedBy:
          type: string
        performedByRole:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        timestamp:
          type: string

    # ── Activity Feed Schemas ──────────────────────────────────────────────────
    ActivityEvent:
      type: object
      required: [id, type, message, timestamp]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [license_assigned, seat_released, user_invited, role_changed, threshold_crossed, campaign_updated, seat_revoked]
        message:
          type: string
        actor:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        severity:
          type: string
          enum: [info, warning, error, success]
        timestamp:
          type: string