# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * B2B AI Marketing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLogEntityType } from './auditLogEntityType';

export interface AuditLog {
  id: number;
  action: string;
  entityType: AuditLogEntityType;
  entityId: number;
  /** @nullable */
  entityName?: string | null;
  performedBy: string;
  /** @nullable */
  performedByRole?: string | null;
  /** @nullable */
  details?: string | null;
  timestamp: string;
}